% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{barn}
\alias{barn}
\alias{print.barn}
\title{Combine datasets to preprocess}
\usage{
barn(..., nominal_sufix = "_cat", numeric_sufix = "_num")

\method{print}{barn}(x, form_width = 30, ...)
}
\arguments{
\item{...}{Extra arguments.}

\item{nominal_sufix}{An optional string for dealing with nominal variables. Defaults to "_cat".}

\item{numeric_sufix}{An optional string for dealing with numeric variables. Defaults to "_num".}

\item{x}{An object of class "barn".}

\item{form_width}{An integer specifying the minimum column width (in characters). Default is 30.}
}
\value{
A barn object containing the combined data frame, row counts,
}
\description{
Combine multiple data frames based on their common columns.
That's the first step for preprocessing with the \code{barn} package.
When printing it shows the characteristics of the combined datasets.
}
\examples{
full <- data.frame(id = 1:3, p1 = c("A", "B", "C"), p2 = 10:12, y = 1:3)
holdout <- data.frame(id = 4:5, p1 = c("D", "E"), p2 = 1:2)
original <- data.frame(id = 1:2, p1 = c("F", "G"), p2 = 3:4, y = 4:5)
print(barn(full, holdout, original))

}
