% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{harvest}
\alias{harvest}
\title{Split the combined dataset}
\usage{
harvest(barn_obj)
}
\arguments{
\item{barn_obj}{An object of class \code{"barn"}, created by \code{\link[=barn]{barn()}}.}
}
\value{
A named list of data frames, one for each dataset originally passed to
\code{\link[=barn]{barn()}}.
}
\description{
Splits the \code{combined} data frame from a \code{barn} object back into a named list
containing the preprocessed predictors.
}
\examples{
full <- data.frame(id = 1:3, p1 = c("A", "B", "C"), p2 = 10:12, y = 1:3)
holdout <- data.frame(id = 4:5, p1 = c("D", "E"), p2 = 1:2)
original <- data.frame(id = 1:2, p1 = c("F", "G"), p2 = 3:4, y = 4:5)
harvested <- barn(full, holdout) |> harvest()
names(harvested)
harvested[["full"]]

}
