% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{plant_count_encode}
\alias{plant_count_encode}
\title{Encode categorical columns with counts}
\usage{
plant_count_encode(barn_obj, nominal_suffix = "_cat")
}
\arguments{
\item{barn_obj}{A Barn object, typically from the \code{bbarchart} package.}

\item{nominal_suffix}{The suffix applied to column names. Defaults to "_cat".}
}
\value{
Modified Barn object with encoded count columns for categorical variables.
}
\description{
Frequency encoding of nominal variables.
}
