% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{plant_new_nominal_pairs}
\alias{plant_new_nominal_pairs}
\title{Create new nominal pairs}
\usage{
plant_new_nominal_pairs(barn_obj, nominal_suffix = "_cat")
}
\arguments{
\item{barn_obj}{An object inheriting from the "barn" class.}

\item{nominal_suffix}{A character string that specifies the suffix for the newly created columns. Optional, default is "_cat".}
}
\value{
The updated barn object with new nominal pairs.
}
\description{
A function to create new features based on combinations of categorical columns in a barn object.
}
