% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{plant_new_numeric_factors}
\alias{plant_new_numeric_factors}
\title{New factors from numerical columns}
\usage{
plant_new_numeric_factors(
  barn_obj,
  numeric_suffx = "_num",
  nominal_suffix = "_cat"
)
}
\arguments{
\item{barn_obj}{A \code{barn} object inheriting from \code{class("barn")}.}

\item{numeric_suffx}{The suffix for new numeric factor columns. Default is "_num".}

\item{nominal_suffix}{The suffix for new nominal factor columns. Default is "_cat".}
}
\value{
The modified \code{barn_obj} with the transformed combined data frame.
}
\description{
A function to transform numeric and character columns in a \code{barn} object into new factor columns.
It appends "_num" for numeric columns, "_cat" for character columns, and renames both to factors.
Original columns are deleted from the combined data frame within the barn object.
}
