% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{plant_summarize}
\alias{plant_summarize}
\title{Summarize Barn Object}
\usage{
plant_summarize(barn_obj, .by = NULL, ...)
}
\arguments{
\item{barn_obj}{An object of class 'barn'.}

\item{.by}{Varible(s) to group by. Currently unused; must be empty.}

\item{...}{Expressions to compute summarizing values.}
}
\value{
A modified \code{barn_obj} with summarized data in the combined slot.
}
\description{
A function to group and summarize to add aggregations to a \code{barn_obj} using specified variables and expressions.
WARNING: Risk of overfitting and bad generalization if not done
when resampling.
}
