% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{plant_count_encode}
\alias{plant_count_encode}
\title{Encode categorical columns with counts}
\usage{
plant_count_encode(barn_obj, nominal_suffix = "_cat")
}
\arguments{
\item{barn_obj}{A Barn object, created by \code{\link[=barn]{barn()}}.}

\item{nominal_suffix}{The suffix applied to column names. Defaults to "_cat".}
}
\value{
The modified \code{barn_obj} with the transformed combined data frame.
}
\description{
Frequency encoding of nominal variables.
}
