% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barn.R
\name{plant_label_encode}
\alias{plant_label_encode}
\title{Encode labels in a barn object}
\usage{
plant_label_encode(barn_obj)
}
\arguments{
\item{barn_obj}{An instance of class "barn".}
}
\value{
The modified \code{barn_obj} with the transformed combined data frame.
}
\description{
Transform nominal columns from factors to integers.
}
